COMPRESS TEST 1

+++ resources/views/items.blade.php
@@ -120,6 +120,10 @@
             </div>
         </div>
         <div class="fixed right-9 bottom-9">
+            <button onclick="getitem()" type="button" class="text-white bg-grey-700 hover:bg-grey-800 focus:ring-4 focus:outline-none focus:ring-grey-300 font-medium rounded-full text-sm p-2.5 text-center inline-flex items-center me-2 dark:bg-grey-600 dark:hover:bg-grey-700 dark:focus:ring-grey-800">
+                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" style="fill: rgba(255, 255, 255, 1);transform: ;msFilter:;"><path d="M12 16c1.671 0 3-1.331 3-3s-1.329-3-3-3-3 1.331-3 3 1.329 3 3 3z"></path><path d="M20.817 11.186a8.94 8.94 0 0 0-1.355-3.219 9.053 9.053 0 0 0-2.43-2.43 8.95 8.95 0 0 0-3.219-1.355 9.028 9.028 0 0 0-1.838-.18V2L8 5l3.975 3V6.002c.484-.002.968.044 1.435.14a6.961 6.961 0 0 1 2.502 1.053 7.005 7.005 0 0 1 1.892 1.892A6.967 6.967 0 0 1 19 13a7.032 7.032 0 0 1-.55 2.725 7.11 7.11 0 0 1-.644 1.188 7.2 7.2 0 0 1-.858 1.039 7.028 7.028 0 0 1-3.536 1.907 7.13 7.13 0 0 1-2.822 0 6.961 6.961 0 0 1-2.503-1.054 7.002 7.002 0 0 1-1.89-1.89A6.996 6.996 0 0 1 5 13H3a9.02 9.02 0 0 0 1.539 5.034 9.096 9.096 0 0 0 2.428 2.428A8.95 8.95 0 0 0 12 22a9.09 9.09 0 0 0 1.814-.183 9.014 9.014 0 0 0 3.218-1.355 8.886 8.886 0 0 0 1.331-1.099 9.228 9.228 0 0 0 1.1-1.332A8.952 8.952 0 0 0 21 13a9.09 9.09 0 0 0-.183-1.814z"></path></svg>
+                <span class="sr-only">REFRESH</span>
+            </button>
             <button onclick="$(window).scrollTop(0);" type="button" class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-full text-sm p-2.5 text-center inline-flex items-center me-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
                 <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" style="fill: rgba(255, 255, 255, 1);transform: ;msFilter:;"><path d="m6.293 13.293 1.414 1.414L12 10.414l4.293 4.293 1.414-1.414L12 7.586z"></path></svg>
                 <span class="sr-only">UP</span>
@@ -157,10 +161,20 @@
                 </div>
             </div>
         </div>
+        {{-- IMG COMPRESSOR START --}}
+        <script src="https://cdn.jsdelivr.net/npm/@sitelintcode/optimize-image-on-the-client-side@0.0.44/dist/optimize-image-on-the-client-side.js"></script>
+        <script>
+            (function() {
+                const optimizeImage = new window.sitelint.OptimizeImage();
+                optimizeImage.install();
+            }())
+        </script>
+        {{-- IMG COMPRESSOR END --}}
         <script>
             var item_properties = {};
             var item_data = [];
             var form_display = false;
+            var ajaxrunnning = 0;
             $(document).ready(function () {
                 $(".slc2").select2({
                     tags: true,
@@ -182,10 +196,16 @@
 
             function getitem() {
                 $.modal.close();
+                if (ajaxrunnning==1) {
+                    alert('Ajax is running. Try again later.....')
+                    return;
+                }
+                ajaxrunnning = 1;
                 $.ajax({
                     type: "GET",
                     url: "/item-data",
                     success: function (response) {
+                        ajaxrunnning = 0;
                         item_data = response.item;
                         var c = '<option></option>';
                         response.category.forEach(e => {
@@ -202,6 +222,10 @@
                             theme: 'tailwindcss-3',
                         });
                         drawitemlistcard();
+                    },
+                    error: function(XMLHttpRequest, textStatus, errorThrown, ) {
+                        ajaxrunnning = 0;
+                        alert(XMLHttpRequest.status + ': ' + (XMLHttpRequest.responseText).replace(/\\/g,''));
                     }
                 });
             }
@@ -374,7 +398,11 @@
                 for (var i = 0; i < files.length; i++) {
                     frmdat.append("images" + i, files[i]);
                 }
-                
+                if (ajaxrunnning==1) {
+                    alert('Ajax is running. Try again later.....')
+                    return;
+                }
+                ajaxrunnning = 1;
                 $.ajax({
                     type: "POST",
                     url: "/item-data",
@@ -382,14 +410,24 @@
                     processData: false,
                     contentType: false,
                     success: function (response) {
+                        ajaxrunnning = 0;
                         resetval();
                         getitem();
+                    },
+                    error: function(XMLHttpRequest, textStatus, errorThrown, ) {
+                        ajaxrunnning = 0;
+                        alert(XMLHttpRequest.status + ': ' + (XMLHttpRequest.responseText).replace(/\\/g,''));
                     }
                 });
             }
 
             function delitemdata(id) {
                 if (confirm('DELETE ITEM?')==true) {
+                    if (ajaxrunnning==1) {
+                        alert('Ajax is running. Try again later.....')
+                        return;
+                    }
+                    ajaxrunnning = 1;
                     $.ajax({
                         type: "DELETE",
                         url: "/item-data/"+id,
@@ -397,8 +435,13 @@
                             _token: $('input[name="_token"]').val(),
                         },
                         success: function (response) {
+                            ajaxrunnning = 0;
                             resetval();
                             getitem();
+                        },
+                        error: function(XMLHttpRequest, textStatus, errorThrown, ) {
+                            ajaxrunnning = 0;
+                            alert(XMLHttpRequest.status + ': ' + (XMLHttpRequest.responseText).replace(/\\/g,''));
                         }
                     });
                 }
@@ -406,6 +449,11 @@
 
             function delimgdata(img,id) {
                 if (confirm('DELETE IMAGE?')==true) {
+                    if (ajaxrunnning==1) {
+                        alert('Ajax is running. Try again later.....')
+                        return;
+                    }
+                    ajaxrunnning = 1;
                     $.ajax({
                         type: "DELETE",
                         url: "/item-data/"+id,
@@ -415,11 +463,20 @@
                             imgurl: img,
                         },
                         success: function (response) {
+                            ajaxrunnning = 0;
                             resetval();
                             getitem();
+                        },
+                        error: function(XMLHttpRequest, textStatus, errorThrown, ) {
+                            ajaxrunnning = 0;
+                            alert(XMLHttpRequest.status + ': ' + (XMLHttpRequest.responseText).replace(/\\/g,''));
                         }
                     });
                 }
+            }
+
+            function compressimage(img) {
+                
             }
 
             function edtitemdata(i) {
